requirements.txt
firebase-admin==6.2.0
python-dotenv==1.0.0
cloudinary==1.40.0
django==4.2.7